refactor(nuekit): simplify, optimise clean up structure
also dropped the custom `subtitle` (a.k.a. "description") option here, and omitted rendering, because it has the same inheritance problem as described in the previous commit for the title. For now, I believe it's the most sensible thing to just ignore the optional (see spec) `subtitle`. Maybe someone can come up with a GOAT solution later.
