+++ packages/examples/nested-blog/blog/child-2/child-2.yaml
@@ -1,5 +1,3 @@
 
 include: [ content, cards, motion ]
 has_feed: true
-feed_title: Something
-feed_description: foo
