+++ packages/nuekit/src/nuekit.js
@@ -13,7 +13,7 @@ import { fswatch } from './nuefs.js'
 
 import { log, colors, getAppDir, parsePathParts, extendData, toPosix } from './util.js'
 import { renderPage, getSPALayout } from './layout/page.js'
-import { getLayoutComponents, generateCollectionFeeds } from './layout/components.js'
+import { getLayoutComponents, collectionToFeed } from './layout/components.js'
 
 
 // the HTML5 doctype (can/prefer lowercase for consistency)
@@ -111,7 +111,7 @@ export async function createKit(args) {
       const key = data.collection_name || cdir
       data[key] = await site.getContentCollection(cdir)
 
-      await generateCollectionFeeds(data, site, write)
+      await generateCollectionFeeds(data, site)
     }
 
     // scripts & styling
@@ -123,6 +123,50 @@ export async function createKit(args) {
     await setupStyles(asset_dir, data)
 
     return { ...data, ...parsePathParts(path), document }
+  }
+
+  // Generate feeds for all `has_feed: true` collections
+  // and their non-excluded subcategories.
+  async function generateCollectionFeeds(data, site) {
+
+    const key = data.collection_name || data.content_collection
+
+    // Group items by dir to later only
+    // need a single call to `getData`.
+
+    const items_by_dir = {}
+    data[key].forEach(item => {
+      (items_by_dir[item.dir] ||= []).push(item)
+    })
+
+    for (const [dir, items] of Object.entries(items_by_dir)) {
+
+      const dir_data = await site.getData(dir)
+
+      // Will be true if explicitly in the collections .yaml, or if
+      // the .yaml in a child directory of a "feedable" parent has
+      // no `has_feed` defined. Excluded when collection or child
+      // explicitly opt-out via `has_feed: false`.
+
+      if (dir_data.has_feed !== true) {
+        try {
+
+          // todo: do we need this or is it a Nue bug?
+          // Nue does not seem to wipe the build folder on new builds.
+          // This can cause leftover `feed.xml` files when the config
+          // in `.yaml` files changes. Hence, we cleanup ourselves.
+
+          const { promises: fs } = await import('node:fs')
+          await fs.unlink(join(site.dist, dir, 'feed.xml'))
+
+        } catch (e) {
+          // No file, all good.
+        }
+        continue
+      }
+
+      await write(...collectionToFeed({ dir, ...dir_data, items }))
+    }
   }
 
   // Markdown page
